Quick fix in ft_humanize_size to fix a compiler overflow error on Linux Mint 32

+++ sources/extra/ft_humanize_size.c
@@ -96,7 +96,7 @@ static void	transform_size_into_str(off_t size, long divider,
 char		*ft_humanize_size(off_t size)
 {
 	char		*str;
-	long		divider;
+	long long	divider;
 	int			i;
 
 	i = 0;
